+++ packages/plugin-hardhat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+- Update Defender SDK. ([#924](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/924))
+- Throw error if not using a relayer for deployments, until other types of deployments are supported.
+
+**Note**: OpenZeppelin Defender deployments is in beta and its functionality is subject to change.
+
 ## 2.4.0 (2023-11-13)
 
 - Add `createFactoryAddress` option for Defender deployments. ([#920](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/920))